% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LPN_survey.R
\name{get_agglevel_names.landholder.plot.neighbour}
\alias{get_agglevel_names.landholder.plot.neighbour}
\title{Aggregation level names}
\usage{
\method{get_agglevel_names}{landholder.plot.neighbour}(object, level)
}
\arguments{
\item{object}{A \code{\link{survey}} object of landholder.plot.neighbour
  subclass.}

\item{level}{A positive integer of length 1.}
}
\value{
A character vector of unique names.
}
\description{
\code{nlandholders} gets all place names which refer to a particular level of
spatial aggregation.
}
\details{
This method extracts a collection of distinct place names from the raw
hyper-table in the survey object.
}
\examples{
get_agglevel_names(tagarina_croplands, 1)
}

