% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LPN_survey.R
\name{get_landholders.landholder.plot.neighbour}
\alias{get_landholders.landholder.plot.neighbour}
\title{Landholders listing}
\usage{
\method{get_landholders}{landholder.plot.neighbour}(object, with_names = TRUE)
}
\arguments{
\item{object}{A \code{\link{survey}} object of landholder.plot.neighbour
  subclass.}

\item{with_names}{A logical flag to get landholder names if TRUE, or
  landholder identifiers otherwise.}
}
\value{
A character vector of unique names or identifiers.
}
\description{
\code{get_landholders} retrieves all landholders in the survey object.
}
\details{
This method extracts the listing of unique landholder names or identifiers
from the raw hyper-table in the survey object.
}
\examples{
get_landholders(tagarina_croplands)
get_landholders(tagarina_croplands, with_names = FALSE)
}

