% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survey.R
\name{get_recordcount.survey}
\alias{get_recordcount.survey}
\title{Survey record count}
\usage{
\method{get_recordcount}{survey}(object)
}
\arguments{
\item{object}{A \code{\link{survey}} object.}
}
\value{
An integer of length 1 or NULL.
}
\description{
\code{get_records} gets number of records in the survey object provided.
}
\details{
This method counts the total number of records in a survey object's
hyper-table.
}
\examples{
get_recordcount(tagarina_croplands)
}

