% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survey.R
\name{get_records.survey}
\alias{get_records.survey}
\title{Survey data retrieval}
\usage{
\method{get_records}{survey}(object)
}
\arguments{
\item{object}{A \code{\link{survey}} object.}
}
\value{
A data frame organized as hyper-table or column-wide store.
}
\description{
\code{get_records} retrieves all records in the survey object provided.
}
\details{
This is the accessor method for the raw hyper-table that holds all records in
a survey object.
}
\examples{
get_records(tagarina_croplands)
}

