% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LPN_survey.R
\name{nagglevels.landholder.plot.neighbour}
\alias{nagglevels.landholder.plot.neighbour}
\title{Aggregation levels count}
\usage{
\method{nagglevels}{landholder.plot.neighbour}(object)
}
\arguments{
\item{object}{A \code{\link{survey}} object of landholder.plot.neighbour
  subclass.}
}
\value{
A non negative integer of length 1.
}
\description{
\code{nlandholders} gets the number of levels land lots can be aggregated in.
}
\details{
This method extracts the number of land lot spatial aggregation levels from
the metadata in the survey object's header.
}
\examples{
nagglevels(tagarina_croplands)
}

