% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LPN_survey.R
\name{nplots.landholder.plot.neighbour}
\alias{nplots.landholder.plot.neighbour}
\title{Plot count}
\usage{
\method{nplots}{landholder.plot.neighbour}(object)
}
\arguments{
\item{object}{A \code{\link{survey}} object of landholder.plot.neighbour
  subclass.}
}
\value{
A non negative integer of length 1.
}
\description{
\code{nlandholders} gets the number of land lots in the survey object.
}
\details{
This method extracts the number of unique land lots from the raw
hyper-table in the survey object.
}
\examples{
nplots(tagarina_croplands)
}

