% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survey.R
\name{survey}
\alias{survey}
\title{Surveys}
\usage{
survey(filename)
}
\arguments{
\item{filename}{Path to a structured text document in one of the supported
formats.}
}
\value{
A survey object or an error in case path is invalid or format/schema
 is not supported.
}
\description{
\code{survey} is a constructor function that generates an instance of a survey
object from an structured text document that collects spatially related
observations.
}
\details{
A survey is a collection of spatially related observations or records that
conform to a structure or schema (e.g. landholder-plot-neighbour) for which
the settle package provides a suitable reader. A survey object is useful to
explore spatial semantics within the collection in order to spatially
aggreggate data, plot neighbouring items or identify and locate clusters.
}
\section{Supported text formats}{
 Currently the only supported format is YAML.
}

\section{Supported schemas}{
 The only implemented reader is LPNschema_reader,
 which can parse documents with a landholder-plot-neighbour schema.
}

